BE: ADded "chat_ws_test_blck".
